requirements.txt
fastmcp>=0.1.0
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0
Pillow>=10.0.0